修改bbc-h5提交订单接口

+++ gs-biz-support/gs-commodity/gs-commodity-repository/src/main/java/com/gs/lshly/biz/support/commodity/entity/GoodsInfo.java
@@ -201,7 +201,7 @@ public class GoodsInfo extends Model {
     /**
      * 积分价格
      */
-    private Double pointPrice;
+    private BigDecimal pointPrice;
     /**
      * 办理备注
      */
